fix tidy

+++ tasks/mpi/kolodkin_g_image_contrast/src/ops_mpi.cpp
@@ -44,8 +44,8 @@ bool kolodkin_g_image_contrast_mpi::TestMPITaskSequential::run() {
   }
   av_br /= input_.size() / 3;
   double k = 1.5;
-  for (int i = 0; i < 256; i++) {
-    int delta_color = (int)i - av_br;
+  for (unsigned long i = 0; i < 256; i++) {
+    unsigned long delta_color = (unsigned long)i - av_br;
     int temp = (int)(av_br + k * delta_color);
     palette_[i] = std::clamp(temp, 0, 255);
   }
